requirements.txt
# Core dependencies
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.1
google-cloud-documentai==2.21.1
google-cloud-aiplatform==1.42.0
google-generativeai==0.3.2

# Configuration and utilities
pydantic==2.6.1
python-dotenv==1.0.1
typing-extensions==4.9.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0

# Development
black==24.1.1
isort==5.13.2
mypy==1.8.0
flake8==7.0.0 